+++ bootstrap/pipe.php
@@ -9,12 +9,12 @@ declare(strict_types=1);
 
 namespace Ixocreate\Framework;
 
-use Ixocreate\Application\Http\ErrorHandling\Response\NotFoundHandler;
 use Ixocreate\Application\Http\Middleware\RootRequestWrapperMiddleware;
 use Ixocreate\Application\Http\Pipe\PipeConfigurator;
 use Ixocreate\Application\Uri\Middleware\ApplicationUriCheckMiddleware;
+use Ixocreate\Framework\Http\ErrorHandling\Response\NotFoundHandler;
 use Ixocreate\Template\Middleware\TemplateMiddleware;
-use Zend\Stratigility\Middleware\ErrorHandler;
+use Laminas\Stratigility\Middleware\ErrorHandler;
 
 /** @var PipeConfigurator $pipe */
 $pipe->pipe(ErrorHandler::class, PHP_INT_MAX);