+++ bootstrap/servicemanager.php
@@ -9,9 +9,9 @@ declare(strict_types=1);
 
 namespace Ixocreate\Framework;
 
-use Ixocreate\Application\Http\ErrorHandling\Factory\ErrorResponseGeneratorFactory;
-use Ixocreate\Application\Http\ErrorHandling\Response\ErrorResponseGenerator;
-use Ixocreate\Application\Service\ServiceManagerConfigurator;
+use Ixocreate\Application\ServiceManager\ServiceManagerConfigurator;
+use Ixocreate\Framework\Http\ErrorHandling\Factory\ErrorResponseGeneratorFactory;
+use Ixocreate\Framework\Http\ErrorHandling\Response\ErrorResponseGenerator;
 
 /** @var ServiceManagerConfigurator $serviceManager */
 $serviceManager->addFactory(ErrorResponseGenerator::class, ErrorResponseGeneratorFactory::class);