+++ src/Package.php
@@ -9,68 +9,10 @@ declare(strict_types=1);
 
 namespace Ixocreate\Framework;
 
-use Ixocreate\Application\Configurator\ConfiguratorRegistryInterface;
-use Ixocreate\Application\Console\ConsoleBootstrapItem;
-use Ixocreate\Application\Console\ConsoleRunner;
-use Ixocreate\Application\Console\ConsoleSubManager;
-use Ixocreate\Application\Console\Factory\ConsoleRunnerFactory;
-use Ixocreate\Application\Http\Factory\FastRouterFactory;
-use Ixocreate\Application\Http\Factory\RequestHandlerRunnerFactory;
-use Ixocreate\Application\Http\Middleware\MiddlewareBootstrapItem;
-use Ixocreate\Application\Http\Middleware\MiddlewareSubManager;
-use Ixocreate\Application\Http\Pipe\PipeBootstrapItem;
 use Ixocreate\Application\Package\PackageInterface;
-use Ixocreate\Application\Publish\PublishBootstrapItem;
-use Ixocreate\Application\Service\ServiceManagerBootstrapItem;
-use Ixocreate\Application\Service\ServiceManagerConfigurator;
-use Ixocreate\Application\Service\ServiceRegistryInterface;
-use Ixocreate\Application\Uri\ApplicationUriBootstrapItem;
-use Ixocreate\ServiceManager\ServiceManagerInterface;
-use Zend\Expressive\Router\FastRouteRouter;
-use Zend\HttpHandlerRunner\RequestHandlerRunner;
 
 final class Package implements PackageInterface
 {
-    /**
-     * @param ConfiguratorRegistryInterface $configuratorRegistry
-     */
-    public function configure(ConfiguratorRegistryInterface $configuratorRegistry): void
-    {
-        /** @var ServiceManagerConfigurator $serviceManagerConfigurator */
-        $serviceManagerConfigurator = $configuratorRegistry->get(ServiceManagerBootstrapItem::class);
-
-        $serviceManagerConfigurator->addFactory(RequestHandlerRunner::class, RequestHandlerRunnerFactory::class);
-        $serviceManagerConfigurator->addFactory(FastRouteRouter::class, FastRouterFactory::class);
-        $serviceManagerConfigurator->addSubManager(MiddlewareSubManager::class);
-
-        $serviceManagerConfigurator->addFactory(ConsoleRunner::class, ConsoleRunnerFactory::class);
-        $serviceManagerConfigurator->addSubManager(ConsoleSubManager::class);
-    }
-
-    /**
-     * @param ServiceRegistryInterface $serviceRegistry
-     */
-    public function addServices(ServiceRegistryInterface $serviceRegistry): void
-    {
-    }
-
-    /**
-     * @return array|null
-     */
-    public function getConfigProvider(): ?array
-    {
-        return null;
-    }
-
-    /**
-     * @param ServiceManagerInterface $serviceManager
-     * @throws \Psr\Container\ContainerExceptionInterface
-     * @throws \Psr\Container\NotFoundExceptionInterface
-     */
-    public function boot(ServiceManagerInterface $serviceManager): void
-    {
-    }
-
     /**
      * @return null|string
      */
@@ -80,35 +22,17 @@ final class Package implements PackageInterface
     }
 
     /**
-     * @return null|string
+     * @return array
      */
-    public function getConfigDirectory(): ?string
+    public function getBootstrapItems(): array
     {
-        return null;
+        return [];
     }
 
     /**
-     * @return array|null
+     * @return array
      */
-    public function getBootstrapItems(): ?array
-    {
-        /**
-         * register the application's bootstrap items
-         * TODO: the application should probably do that by itself
-         */
-        return [
-            ApplicationUriBootstrapItem::class,
-            ConsoleBootstrapItem::class,
-            MiddlewareBootstrapItem::class,
-            PipeBootstrapItem::class,
-            PublishBootstrapItem::class,
-        ];
-    }
-
-    /**
-     * @return array|null
-     */
-    public function getDependencies(): ?array
+    public function getDependencies(): array
     {
         /**
          * require all the framework's default packages