+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Release Notes
 
-## [Unreleased](https://github.com/ixocreate/framework/compare/0.3.2...develop)
+## [Unreleased](https://github.com/ixocreate/framework/compare/0.4.1...develop)
+
+## [v0.3.3 (2019-09-18)](https://github.com/ixocreate/framework/compare/0.3.2...0.3.3)
+### Fixed
+- Fix DatabasePrepareConsole command
+
+## [v0.4.1 (2019-09-13)](https://github.com/ixocreate/framework/compare/0.4.0...0.4.1)
+### Fixed
+- Fix DatabasePrepareConsole command
 
 ## [v0.4.0 (2019-07-17)](https://github.com/ixocreate/framework/compare/0.3.2...0.4.0)
 ### Changed
-- Upgrade to Media Package v0.4:
+- Upgrade to Media Package v0.4
 
 ## [v0.3.2 (2019-06-27)](https://github.com/ixocreate/framework/compare/0.3.1...0.3.2)
 ### Fixed